This version is based on Citezen-tg.45.
Added BibList and BibEntry structures. The BibList takes as sources both a local file and a list of urls to bib files. When the sources are set, BibEntries are created as subpages. Thus, we can refer to bib entries (*biblist/bibentrykey*) or we can directly embed a paper description (+biblist/bibentrykey+). 

Furthermore, the keywords from the bib entry are added as tags of the BibEntry. Like this, we can use the TagCloud to display the keywords of the papers.